Refactor FlowScanner & affiliated classes to be prepare for review
